+++ sample/sample-with-di/app/src/main/java/com/dropbox/kaiken/sample/advancedsample/app/LoginActivity.kt
@@ -29,6 +29,7 @@ fun getLaunchIntent(
     userId: String,
 ): Intent = Intent(context, HomeActivity::class.java).apply {
     putViewingUserSelector(ViewingUserSelector.fromUserId(userId))
+
 }
 
 @ContributesTo(AuthOptionalScreenScope::class)
