+++ sample/sample-with-di/helloworldfeature/src/main/java/com/dropbox/kaiken/sample/advancedsample/helloworldfeature/LoginRouter.kt
@@ -3,14 +3,16 @@ package com.dropbox.kaiken.sample.advancedsample.helloworldfeature
 import android.content.Context
 import android.content.Intent
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.NavController
+import androidx.navigation.NavBackStackEntry
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
 import com.dropbox.common.inject.AuthOptionalScreenScope
 import com.dropbox.kaiken.Injector
 import com.dropbox.kaiken.skeleton.scoping.cast
 import com.squareup.anvil.annotations.ContributesTo
+import kotlinx.coroutines.flow.collect
 
 @ContributesTo(AuthOptionalScreenScope::class)
 interface LoginScreenComponent : Injector {
@@ -21,15 +23,21 @@ interface LoginScreenComponent : Injector {
 fun LoginRouter() {
     val navController = rememberNavController()
     NavHost(navController = navController, startDestination = "login") {
-        authAwareComposable("login") { _, presenter: LoginPresenter ->
+        authAwareComposable("login") { entry: NavBackStackEntry, presenter: LoginPresenter ->
+            LaunchedEffect(presenter.effects) {
+                presenter.effects.collect {
+                    when(it) {
+                        is LoginPresenter.LoginSuccessful -> {
+                            val intentFactory = this@authAwareComposable.cast<LoginScreenComponent>().intentFactory()
+                            navController.context.startActivity(intentFactory(navController.context, it.userId))
+                        }
+                    }
+                }
+            }
+
             LoginScreen(
                 presenter.model,
-                { submit: LoginPresenter.Submit -> presenter.events.tryEmit(submit) },
-                { userId: String ->
-                    val intentFactory = this.cast<LoginScreenComponent>().intentFactory()
-                    navController.context.startActivity(intentFactory(navController.context, userId))
-
-                }
+                { submit: LoginPresenter.Submit -> presenter.events.tryEmit(submit) }
             ) { navController.navigate("forgot_password") }
         }
         authAwareComposable("forgot_password") { _, presenter: ForgotPasswordPresenter ->
@@ -43,7 +51,7 @@ fun LoginRouter() {
 @Preview
 @Composable
 fun previewLoginScreen() {
-    LoginScreen(model = LoginPresenter.LoginNeeded, onSubmit = {}, onLoggedInSuccess = {}) {}
+    LoginScreen(model = LoginPresenter.LoginNeeded, onSubmit = {}) {}
 }
 
 @Preview
