+++ sample/sample-with-di/helloworldfeature/src/main/java/com/dropbox/kaiken/sample/advancedsample/helloworldfeature/LoginScreen.kt
@@ -1,7 +1,5 @@
 package com.dropbox.kaiken.sample.advancedsample.helloworldfeature
 
-import android.content.Context
-import android.content.Intent
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material.Button
@@ -9,13 +7,11 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.material.TextField
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import com.dropbox.common.inject.AuthOptionalScreenScope
 import com.dropbox.kaiken.skeleton.core.SkeletonOauth2
 import com.dropbox.kaiken.skeleton.scoping.SingleIn
@@ -24,13 +20,15 @@ import javax.inject.Inject
 
 class UserInput(val userId: String)
 abstract class LoginPresenter :
-    Presenter<LoginPresenter.LoginEvent, LoginPresenter.LoginModel>(LoginNeeded) {
+    Presenter<LoginPresenter.LoginEvent, LoginPresenter.LoginModel, LoginPresenter.LoginEffect>(LoginNeeded) {
     sealed interface LoginEvent
     data class Submit(val userInput: UserInput) : LoginEvent
 
     sealed interface LoginModel
-    data class LoginSuccess(val userId: String) : LoginModel
     object LoginNeeded : LoginModel
+
+    sealed interface LoginEffect
+    data class LoginSuccessful(val userId: String) : LoginEffect
 }
 
 @SingleIn(AuthOptionalScreenScope::class)
@@ -52,7 +50,7 @@ class RealLoginPresenter @Inject constructor(
                         "Bart Always Bart"
                     )
                 )
-                model = LoginSuccess(userId = event.userInput.userId)
+                emitEffect(LoginSuccessful(userId = event.userInput.userId))
             }
         }
     }
@@ -62,7 +60,6 @@ class RealLoginPresenter @Inject constructor(
 fun LoginScreen(
     model: LoginPresenter.LoginModel,
     onSubmit: (LoginPresenter.Submit) -> Unit,
-    onLoggedInSuccess: (String) -> Unit,
     onClickForgotPassword: () -> Unit
 ) {
     MaterialTheme {
@@ -70,9 +67,6 @@ fun LoginScreen(
             when (model) {
                 is LoginPresenter.LoginNeeded -> {
                     loginView(onSubmit, onClickForgotPassword)
-                }
-                is LoginPresenter.LoginSuccess -> {
-                    onLoggedInSuccess(model.userId)
                 }
             }
         }
